+++ lib/src/select.dart
@@ -1,19 +1,19 @@
 part of intl_message;
 
 abstract class SubMessage extends ExpressionSubstitution {
-  final Map<String, IntlMessage> /*!*/ messages;
+  final Map<String, IntlMessage> messages;
 
   SubMessage(Expression name, this.messages)
       : super(name, fallbackToNullWhenEvaluationFails: true);
 
-  String /*!*/ _index(covariant dynamic v);
+  String? _index(covariant dynamic v);
 
   String get _type;
 
   @override
   FutureOr<String> formatter(v, Map<String, dynamic> args) {
     var index = _index(v);
-    var m = messages[index] ?? messages['other'];
+    var m = messages[index] ?? messages['other']!;
     return m.format(args);
   }
 
@@ -30,7 +30,7 @@ class SelectMessage extends SubMessage {
       : super(name, messages);
 
   @override
-  String _index(String v) => v;
+  String? _index(String? v) => v;
 
   @override
   String get _type => 'select';
@@ -47,8 +47,8 @@ class SelectOrdinalMessage extends PluralMessage {
     ordinal_rules.startRuleEvaluation(howMany);
     var verifiedLocale = Intl.verifiedLocale(
         locale, ordinal_rules.localeHasPluralRules,
-        onFailure: (locale) => 'default') /*!*/;
-    return ordinal_rules.pluralRules[verifiedLocale] /*!*/ ();
+        onFailure: (locale) => 'default')!;
+    return ordinal_rules.pluralRules[verifiedLocale]!();
   }
 
   @override
@@ -63,8 +63,7 @@ class PluralMessage extends SubMessage {
       : super(name, messages);
 
   @override
-  FutureOr<String> format(Map<String, dynamic> args,
-      {ErrorHandler /*?*/ onError}) {
+  FutureOr<String> format(Map<String, dynamic> args, {ErrorHandler? onError}) {
     var s = super.format(args, onError: onError);
     return _replace(s, args);
   }
@@ -78,9 +77,9 @@ class PluralMessage extends SubMessage {
   FutureOr<String> _replace(FutureOr<String> s, v) {
     if (s is String) {
       return s.replaceAllMapped(RegExp(r'(^|[^\\])#'),
-          (m) => m.group(1) + NumberFormat().format(v - offset));
+          (m) => m.group(1)! + NumberFormat().format(v - offset));
     }
-    return (s as Future<String>).then((s) => _replace(s, v));
+    return s.then((s) => _replace(s, v));
   }
 
   plural_rules.PluralCase _pluralCase(int howMany) {
@@ -88,8 +87,8 @@ class PluralMessage extends SubMessage {
     plural_rules.startRuleEvaluation(howMany);
     var verifiedLocale = Intl.verifiedLocale(
         locale, plural_rules.localeHasPluralRules,
-        onFailure: (locale) => 'default') /*!*/;
-    return plural_rules.pluralRules[verifiedLocale] /*!*/ ();
+        onFailure: (locale) => 'default')!;
+    return plural_rules.pluralRules[verifiedLocale]!();
   }
 
   @override